Update function name and arguments in 07
Noticed that the column checking function in the latter chunks did not match the definition in the first chunk. Not sure if the name should be changed in the first or second chunk. Not sure about the `match.call()` argument either - has it been used in an old version? Anyway the function now seems to produce the kind of an error message that the text refers to.