Update 03_agile-ds.Rmd
Adding "when" feels like it fits better with what I think the intent of the sentence is. Currently, it feels like "business people get involved" and "they have strong opinions" are a list of things that happen (a, then b) but aren't strong linked (a, then b because a happened).